requirements.txt
Flask>=2.0.0
line-bot-sdk>=2.0.0
google-cloud-firestore>=2.5.0
python-dotenv>=0.15.0